Moved platform-specific time types and functions to platform support
