+++ org.lflang/src/lib/core/reactor_common.c
@@ -1786,31 +1786,6 @@ int process_args(int argc, char* argv[]) {
 }
 
 /**
- * Calculate the necessary offset to bring _LF_CLOCK in parity
- * with the epoch time.
- */
-void calculate_epoch_offset() {
-    if (_LF_CLOCK == CLOCK_REALTIME) {
-        // Set the epoch offset to zero (see tag.h)
-        _lf_epoch_offset = 0LL;
-    } else {
-        // Initialize _lf_epoch_offset to the difference between what is
-        // reported by whatever clock LF is using (e.g. CLOCK_MONOTONIC)
-        // and what is reported by CLOCK_REALTIME.
-        struct timespec physical_clock_snapshot, real_time_start;
-
-        clock_gettime(_LF_CLOCK, &physical_clock_snapshot);
-        instant_t physical_clock_snapshot_ns = physical_clock_snapshot.tv_sec * BILLION + physical_clock_snapshot.tv_nsec;
-
-        clock_gettime(CLOCK_REALTIME, &real_time_start);
-        instant_t real_time_start_ns = real_time_start.tv_sec * BILLION + real_time_start.tv_nsec;
-
-        _lf_epoch_offset = real_time_start_ns - physical_clock_snapshot_ns;
-    }
-    LOG_PRINT("Clock sync: Initial epoch offset set to %lld.", _lf_epoch_offset);
-}
-
-/**
  * Initialize the priority queues and set logical time to match
  * physical time. This also prints a message reporting the start time.
  */
